+++ src/03/second_way.sh
@@ -11,13 +11,11 @@ second_way () {
     process_status_bar $start $finish &
     local pid=$!
     for path in ${result[@]}; do
-        if ! rm -rf $path 2>/dev/null; then
-            if ! rmdir $value 2>/dev/null; then
-                echo -e "\n\n${RED}Ошибка${RESET}: Не удается очистить файловую систему"
-                echo "Возможно требуется запустить скрипт с root правами"
-                kill -SIGINT "$pid" ## попробовать без флага
-                return 1
-            fi
+        if ! rm -rf $path 2>/dev/null && ! rmdir $path 2>/dev/null; then
+            echo -e "\n\n${RED}Ошибка${RESET}: Не удается очистить файловую систему"
+            echo "Возможно требуется запустить скрипт с root правами"
+            kill -SIGINT "$pid" ## попробовать без флага
+            return 1
         fi
     done
     wait "$pid"
